fix storage types

+++ internal/app/app.go
@@ -9,8 +9,8 @@ type App struct {
 type Storage interface {
 	CreateTask(string, string) error
 	// UpdateTask(int64, *storage.Task) error
-	DeleteTask(int64) error
-	ShowTask(int64) (storage.Task, error)
+	DeleteTask(int) error
+	ShowTask(int) (storage.Task, error)
 	ShowList() ([]storage.Task, error)
 	Migrate() error
 }
